Reorganized tests to reduce duplication, and running tests for both new and recycled sandboxes.
