+++ test/test.html
@@ -7,9 +7,8 @@
   <script src="dist/abecedary.js"></script>
   <script src="node_modules/mocha/mocha.js"></script>
   <script src="node_modules/chai/chai.js"></script>
-  <script> mocha.setup('bdd');</script>
-  <script src="test/legacy-test.js"></script>
-  <script src="test/systemjs-test.js"></script>
+  <script> mocha.setup({ui: 'bdd', fullTrace: true});</script>
+  <script src="test/tests.js"></script>
   <script>
     window.onload = function() {
         mocha.run()
